No need to explicitly include herr.h anymore

+++ mfhdf/libsrc/local_nc.h
@@ -46,7 +46,6 @@
 
 #include "hdf.h"
 #include "vg.h"
-#include "herr.h"
 
 #define ATTR_TAG  DFTAG_VH
 #define DIM_TAG   DFTAG_VG
@@ -472,8 +471,8 @@ extern int hdf_read_sds_cdf
     PROTO((XDR *,NC **));
 
 extern int NCgenio
-    PROTO((NC *handle,int varid,const int32 *start,const int32*count,
-        const int32 *stride,const int32 *imap,Void *values));
+    PROTO((NC *handle, int varid, const long *start, const long *count,
+        const long *stride, const long *imap,Void *values));
 
 extern int NC_var_shape
     PROTO((NC_var *var,NC_array *dims));
